break up big panel
